tester_api/requirements.txt
#Arquivo tester_api/requirements.txt
Flask
kafka-python
